+++ Plugins/Projects/src/main/java/de/fraunhofer/iosb/ilt/frostserver/plugin/projects/PluginProjects.java
@@ -17,26 +17,44 @@
  */
 package de.fraunhofer.iosb.ilt.frostserver.plugin.projects;
 
+import static de.fraunhofer.iosb.ilt.frostserver.persistence.pgjooq.utils.Utils.getFieldOrNull;
 import static de.fraunhofer.iosb.ilt.frostserver.service.InitResult.INIT_DELAY;
 
+import de.fraunhofer.iosb.ilt.frostserver.model.EntityType;
+import de.fraunhofer.iosb.ilt.frostserver.model.ModelRegistry;
+import de.fraunhofer.iosb.ilt.frostserver.model.core.Entity;
+import de.fraunhofer.iosb.ilt.frostserver.model.core.PkValue;
+import de.fraunhofer.iosb.ilt.frostserver.persistence.PersistenceManager;
+import de.fraunhofer.iosb.ilt.frostserver.persistence.pgjooq.JooqPersistenceManager;
+import de.fraunhofer.iosb.ilt.frostserver.persistence.pgjooq.tables.StaMainTable;
+import de.fraunhofer.iosb.ilt.frostserver.persistence.pgjooq.tables.StaTable;
+import de.fraunhofer.iosb.ilt.frostserver.persistence.pgjooq.tables.TableCollection;
 import de.fraunhofer.iosb.ilt.frostserver.plugin.actuation.PluginActuation;
 import de.fraunhofer.iosb.ilt.frostserver.plugin.coremodel.PluginCoreModel;
+import de.fraunhofer.iosb.ilt.frostserver.plugin.coremodel.TableImpFeatures;
+import de.fraunhofer.iosb.ilt.frostserver.plugin.coremodel.TableImpLocations;
 import de.fraunhofer.iosb.ilt.frostserver.plugin.modelloader.PluginModelLoader;
 import de.fraunhofer.iosb.ilt.frostserver.plugin.multidatastream.PluginMultiDatastream;
 import de.fraunhofer.iosb.ilt.frostserver.service.InitResult;
-import de.fraunhofer.iosb.ilt.frostserver.service.Plugin;
 import de.fraunhofer.iosb.ilt.frostserver.service.PluginManager;
+import de.fraunhofer.iosb.ilt.frostserver.service.PluginModel;
+import de.fraunhofer.iosb.ilt.frostserver.service.UpdateMode;
 import de.fraunhofer.iosb.ilt.frostserver.settings.ConfigDefaults;
 import de.fraunhofer.iosb.ilt.frostserver.settings.CoreSettings;
 import de.fraunhofer.iosb.ilt.frostserver.settings.Settings;
 import de.fraunhofer.iosb.ilt.frostserver.settings.annotation.DefaultValueBoolean;
+import org.jooq.Field;
+import org.jooq.Record1;
+import org.jooq.Record2;
+import org.jooq.Result;
+import org.jooq.TableField;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * Plugin loader for the Projects plugin.
  */
-public class PluginProjects implements Plugin, ConfigDefaults {
+public class PluginProjects implements PluginModel, ConfigDefaults {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(PluginProjects.class.getName());
 
@@ -46,10 +64,18 @@ public class PluginProjects implements Plugin, ConfigDefaults {
     @DefaultValueBoolean(true)
     public static final String TAG_DEFAULT_RULES = "projects.enableDefaultRules";
 
+    @DefaultValueBoolean(false)
+    public static final String TAG_UPDATE_FEATURE_WITH_LOCATION = "projects.updateFeatureWithLocation";
+
+    private CoreSettings coreSettings;
+
     private boolean enabled;
+    private boolean autoUpdateFeatures;
+    private boolean fullyInitialised;
 
     @Override
     public InitResult init(CoreSettings coreSettings) {
+        this.coreSettings = coreSettings;
         Settings pluginSettings = coreSettings.getPluginSettings();
         enabled = pluginSettings.getBoolean(TAG_ENABLE_PROJECTS, PluginProjects.class);
         if (enabled) {
@@ -116,6 +142,8 @@ public class PluginProjects implements Plugin, ConfigDefaults {
                     pml.addSecurityFile("pluginprojects/sta1/security/.json");
                 }
             }
+            autoUpdateFeatures = pluginSettings.getBoolean(TAG_UPDATE_FEATURE_WITH_LOCATION, PluginProjects.class);
+
             pluginManager.registerPlugin(this);
         }
         return InitResult.INIT_OK;
@@ -124,6 +152,111 @@ public class PluginProjects implements Plugin, ConfigDefaults {
     @Override
     public boolean isEnabled() {
         return enabled;
+    }
+
+    @Override
+    public void registerEntityTypes() {
+        // Nothing to do here.
+    }
+
+    @Override
+    public boolean linkEntityTypes(PersistenceManager pm) {
+        if (autoUpdateFeatures) {
+            final PluginCoreModel pluginCoreModel = coreSettings.getPluginManager().getPlugin(PluginCoreModel.class);
+            if (pluginCoreModel == null || !pluginCoreModel.isFullyInitialised()) {
+                return false;
+            }
+            if (pm instanceof JooqPersistenceManager jpm) {
+                StaMainTable<?> tableLocations = jpm.getTableCollection().getTableForType(pluginCoreModel.etLocation);
+                tableLocations.registerHookPostUpdate(
+                        11.1,
+                        (JooqPersistenceManager jpml, Entity entity, PkValue entityId, UpdateMode updateMode) -> relinkProjectsToFeature(jpml, entity));
+            }
+        }
+        fullyInitialised = true;
+        return true;
+    }
+
+    public static void relinkProjectsToFeature(JooqPersistenceManager jpm, Entity entity) {
+        Object genFoiId = null;
+        try {
+            final TableCollection tableCollection = jpm.getTableCollection();
+            final ModelRegistry mr = jpm.getCoreSettings().getModelRegistry();
+
+            EntityType etProject = mr.getEntityTypeForName("Project", true);
+            if (etProject == null) {
+                LOGGER.error("No Project entity type found!");
+                return;
+            }
+            TableImpLocations ql = tableCollection.getTableForClass(TableImpLocations.class);
+            Field<?> lRestricted = ql.field("RESTRICTED");
+            final Object locationId = entity.getPrimaryKeyValues().get(0);
+            Record2<?, ?> fetchAny = jpm.getDslContext()
+                    .select(ql.colGenFoiId, lRestricted)
+                    .from(ql)
+                    .where(((TableField) ql.getId()).equal(locationId))
+                    .fetchAny();
+            if (fetchAny == null) {
+                LOGGER.error("Location not found in table for id {}", entity.getPrimaryKeyValues());
+                return;
+            }
+            genFoiId = fetchAny.component1();
+            if (genFoiId == null) {
+                // Location has no genFoiID
+                return;
+            }
+
+            TableImpFeatures tf = tableCollection.getTableForClass(TableImpFeatures.class);
+            TableField fRestricted = (TableField) tf.field("RESTRICTED");
+            int updated = jpm.getDslContext()
+                    .update(tf)
+                    .set(fRestricted, fetchAny.component2())
+                    .where(((TableField) tf.getId()).eq(genFoiId))
+                    .execute();
+            if (updated != 1) {
+                LOGGER.warn("Update of generated FoI resulted in {} changed rows!", updated);
+            }
+
+            StaTable tfp = tableCollection.getTableForName("FEATURE_PROJECTS");
+            TableField tfpPid = (TableField) tfp.field("PROJECT_ID");
+            TableField tfpFid = (TableField) tfp.field("FEATURE_ID");
+            StaTable tlp = tableCollection.getTableForName("LOCATION_PROJECTS");
+            TableField tlpPid = (TableField) tlp.field("PROJECT_ID");
+            TableField tlpLid = (TableField) tlp.field("LOCATION_ID");
+            if (tfpPid == null || tfpFid == null || tlpPid == null || tlpLid == null) {
+                LOGGER.error("Failed to link Projects to generated FoI, linktables do not have correct columns.");
+                return;
+            }
+            int unlinked = jpm.getDslContext()
+                    .deleteFrom(tfp)
+                    .where(tfpFid.eq(genFoiId))
+                    .execute();
+            LOGGER.debug("Unlinked {} projects from Feature {}", unlinked, genFoiId);
+            Result<Record1<Object>> projectIds = jpm.getDslContext()
+                    .select(tlpPid)
+                    .from(tlp)
+                    .where(tlpLid.eq(locationId))
+                    .fetch();
+            int linked = 0;
+            for (org.jooq.Record pidTuple : projectIds) {
+                Object projectId = getFieldOrNull(pidTuple, tlpPid);
+                jpm.getDslContext()
+                        .insertInto(tfp)
+                        .columns(tfpFid, tfpPid)
+                        .values(genFoiId, projectId)
+                        .execute();
+                linked++;
+            }
+            LOGGER.debug("Linked {} projects from Feature {}", linked, genFoiId);
+
+        } catch (RuntimeException ex) {
+            LOGGER.error("Failed to update Projects for generated FoI {} from Location {}.", genFoiId, entity, ex);
+        }
+    }
+
+    @Override
+    public boolean isFullyInitialised() {
+        return fullyInitialised;
     }
 
 }